Add minimal example and move advanced demo
