+++ example/lib/advanced/geometry/geometry_page.dart
