+++ example/lib/advanced/notifiers/root_notifier.dart
