+++ example/lib/advanced/paiting/border_shapes.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:material_toolkit/material_toolkit.dart';
-import 'package:material_toolkit_example/widgets/group_item_container.dart';
-import 'package:material_toolkit_example/widgets/group_item_row.dart';
-import 'package:material_toolkit_example/widgets/group_item_title.dart';
+import 'package:material_toolkit_example/advanced/widgets/group_item_container.dart';
+import 'package:material_toolkit_example/advanced/widgets/group_item_row.dart';
+import 'package:material_toolkit_example/advanced/widgets/group_item_title.dart';
 
 class BorderRadiusGroup extends StatelessWidget {
   const BorderRadiusGroup({super.key});
