+++ example/lib/advanced/widgets/group_card.dart
