+++ example/lib/advanced/widgets/group_item_container.dart
