+++ example/lib/advanced/widgets/group_item_row.dart
