+++ example/lib/advanced/widgets/info_component.dart
